Issue#33 - crud independent of config.
