+++ app/pillar/handler/config.go
@@ -1,0 +1,67 @@
+package handler
+
+import (
+	"fmt"
+	"log"
+	"os"
+	"strings"
+)
+
+const (
+	LogFile    string = "pillar.log"
+)
+
+//Single logger for the application
+var (
+	Logger *log.Logger
+)
+
+func init() {
+	home := os.Getenv("PILLAR_HOME")
+	if home == "" {
+		log.Fatal("Error initializing Pillar: PILLAR_HOME not found.")
+	}
+
+	url := os.Getenv("MONGODB_URL")
+	if url == "" {
+		log.Fatal("Error initializing Pillar: MONGODB_URL not found.")
+	}
+
+	logFile := getLogFile(home)
+	file, err := os.OpenFile(logFile, os.O_WRONLY|os.O_CREATE|os.O_APPEND, 0666)
+	if err != nil {
+		log.Fatalf("Error opening log file [%s], %s", logFile, err)
+	}
+	fmt.Printf("Pillar log file: %s\n\n", logFile)
+
+	Logger = log.New(file, "Pillar: ", log.LstdFlags|log.Llongfile|log.Ldate|log.Ltime)
+}
+
+func getLogFile(pillarHome string) string {
+	if pillarHome == "" {
+		pillarHome = strings.TrimSuffix(os.TempDir(), "/") + "/pillar"
+	}
+
+	logPath := pillarHome + "/logs"
+	if !exists(logPath) {
+		os.MkdirAll(logPath, 0700)
+	}
+
+	return logPath + "/" + LogFile
+}
+
+func isDir(path string) bool {
+	if info, err := os.Stat(path); err == nil && info.IsDir() {
+		return true
+	}
+
+	return false
+}
+
+func exists(path string) bool {
+	if _, err := os.Stat(path); err == nil {
+		return true
+	}
+
+	return false
+}
