+++ config/localhost.cfg
@@ -1,0 +1,2 @@
+#export PILLAR_HOME=
+#export MONGODB_URL=mongodb://user:password@host:port/database?options
